Debugging plot_t_vs_ev and more documentation
Added more docs for threshold(), but this still needs work.
Especially for the local-global and rank portions of the
parameters.

Fixed a weird bug in plot_t_vs_ev() that would try to plot
a point further than the range of the x-axis.  Tried this
with HumanCellCycleSubset with RMT method. The recommended
threshold was 1.34, but I haven't confirmed whether this is
the same for Carissa's get_results().

Added a temporary check for "hey don't plot this point
if it will be out of the threshold range [t_begin,t_end]".
Will diagnose further after checking against her code.