Error checking get_results()
Fixed error in get_results() helper functions where an empty
file would try and be read. This read takes place in a tryCatch()
block now. Also fixed the Random Matrix Theory bug that did not
properly compute the value in the same way as Carissa's package

IN THE FUTURE  -  need to validate the format of an analysis()
produced output file, not just that there is not an empty file,
which is likely due to analysis failing or ending early.