Documentation and cleaning...
Deleted some junk files from the project root directory. Also
modified a lot of the documentation files to either add
content or provide better examples. Almost done with most of the
examples, just need some for the thresholding functions
themselves.

Placing the sample raw data in "inst/extata" worked as intended
and was able to be found in "extdata" with system.file() after
the package was installed. This is good.