Take 1 at fixing build errors for Viktor (Mac)
Replaced some problem calls to sprintf to snprintf
that Xcode was complaining about in the 'igraph'
library. This involve retarred the modified package.

Preemptively modified some of the alglib and thresholding
package util file calls to sprintf. Hopefully the entire igraph
library doesn't have to make these changes, but the fix would
at least be doable if not tedious.