Removing local-global method
This method took too long to run, and wasn't a part of the
15 methods studied by the 2024 JCB thresholding paper. Removed
dedicated wrapper calls, analysis method support, and updated
documentation to handle the removal of the method.

Some further proof reading and modifications may need to be made,
but this version will build and compile without the local-global
method.