Remove user input error for analysis()
Previously, a use of a static keyword declared vector
would not apply new user inputs to graph files, instead
using the first parameters passed to the function. This will create
a const vector instead.

Maybe could play around with the possibility of creating a static
object initially, then assigning the value of the vector on another
line (decouple the declaration and initialization) of the vector
if the performance is not doing well.